requirements.txt
pytorch
numpy
gym
gym-atari
wandb
gym-box2d